# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Recruit API",
        "description": "採用広報サイト + 応募管理ダッシュボード API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/inquiries": {
            "post": {
                "tags": [
                    "Inquiries"
                ],
                "summary": "問い合わせ送信（公開）",
                "description": "公開側: 問い合わせ送信",
                "operationId": "1cd5d3c54f3754afa8ae1e8c3636ae86",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "送信完了"
                    },
                    "422": {
                        "description": "バリデーションエラー"
                    }
                }
            }
        },
        "/admin/inquiries": {
            "get": {
                "tags": [
                    "Admin/Inquiries"
                ],
                "summary": "問い合わせ一覧（管理）",
                "description": "管理側: 問い合わせ一覧",
                "operationId": "29b72c770c594f17dcc9f507ee2e42e7",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "問い合わせ一覧"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/inquiries/{inquiry}": {
            "get": {
                "tags": [
                    "Admin/Inquiries"
                ],
                "summary": "問い合わせ詳細（管理）",
                "description": "管理側: 問い合わせ詳細",
                "operationId": "ded95ae3a698a92c79e79bf271fa768f",
                "parameters": [
                    {
                        "name": "inquiry",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "問い合わせ詳細"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin/Inquiries"
                ],
                "summary": "問い合わせ更新（管理）",
                "description": "管理側: 対応済み / メモ更新",
                "operationId": "f41a7f7c43cef27e96e6abe779e131c7",
                "parameters": [
                    {
                        "name": "inquiry",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "nullable": true,
                                        "enum": [
                                            "unread",
                                            "replied"
                                        ]
                                    },
                                    "admin_note": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "更新成功"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "求人一覧（公開）",
                "description": "公開側: 求人一覧（published のみ）",
                "operationId": "eda7f0ed621e65cfaceeb96ea64311ab",
                "parameters": [
                    {
                        "name": "employment_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "求人一覧"
                    }
                }
            }
        },
        "/jobs/{slug}": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "求人詳細（公開）",
                "description": "公開側: 求人詳細（slug）",
                "operationId": "4e7c01b403bbd6fa4ca7f5cdbcfcf2bc",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "求人詳細"
                    },
                    "404": {
                        "description": "見つからない"
                    }
                }
            }
        },
        "/admin/jobs": {
            "get": {
                "tags": [
                    "Admin/Jobs"
                ],
                "summary": "求人一覧（管理）",
                "description": "管理側: 求人一覧（全件）",
                "operationId": "8164dfce5fee93415011f63728431ffc",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "求人一覧"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin/Jobs"
                ],
                "summary": "求人作成",
                "description": "管理側: 求人作成",
                "operationId": "5fee0569683e7b5127854a419d344d21",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "作成成功"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/jobs/{job_posting}": {
            "get": {
                "tags": [
                    "Admin/Jobs"
                ],
                "summary": "求人詳細（管理）",
                "description": "管理側: 求人詳細",
                "operationId": "faeb1e0d5f74b2cc20eaff8d1dca25b5",
                "parameters": [
                    {
                        "name": "job_posting",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "求人詳細"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin/Jobs"
                ],
                "summary": "求人更新",
                "description": "管理側: 求人更新",
                "operationId": "1d6783cfc0b0f1d6337ff7bdb30eae53",
                "parameters": [
                    {
                        "name": "job_posting",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "更新成功"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin/Jobs"
                ],
                "summary": "求人削除",
                "description": "管理側: 求人削除",
                "operationId": "a2d974b90f726624ad4b0bdb4b20797c",
                "parameters": [
                    {
                        "name": "job_posting",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "削除成功"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}